contrib/netsurf/libhubbub/test/data/tree-construction/regression.dat

#data
<html><body><table><tr><td><div><b></div>X<table></table></td></tr></table><table></table><script type="text/javascript"></script></body></html>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           <td>
|             <div>
|               <b>
|             <b>
|               "X"
|               <table>
|     <table>
|     <script>
|       type="text/javascript"

#data
<font><center></font>
#errors
#document
| <html>
|   <head>
|   <body>
|     <font>
|     <center>
|       <font>

#data
<a href="a&amp;b&amp;c">Foo</a>
#errors
#document
| <html>
|   <head>
|   <body>
|     <a>
|       href="a&b&c"
|       "Foo"

#data
<table><col width="400"><col><td>width</td><td>nowidth</td></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <colgroup>
|         <col>
|           width="400"
|         <col>
|       <tbody>
|         <tr>
|           <td>
|             "width"
|           <td>
|             "nowidth"

#data
<table><tr bgcolor="yellow"><td>yellow</td></tr><tr><td>transparent</td></tr></table>
#errors
#document
| <html>
|   <head>
|   <body>
|     <table>
|       <tbody>
|         <tr>
|           bgcolor="yellow"
|           <td>
|             "yellow"
|         <tr>
|           <td>
|             "transparent"

#data
<svg xmlns="http://www.w3.org/2000/svg"></svg>
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
|       xmlns xmlns="http://www.w3.org/2000/svg"
